backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["EchoFoe <aspirpd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
channels-redis = "^4.2.0"
redis = "^5.0.1"
daphne = "^4.0.0"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.2"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
